requirements.txt
openai
python-dotenv
google-generativeai
python-magic
